feat: change VForm file of palce
