--- src/components/VForm/VInput/constants.ts
@@ -1,1 +1,0 @@
-export const EXCESS_PROPS = ['rules', 'modelValue', 'errorMessage', 'error', 'bottomSlots'];
