+++ src/shared/ui/VForm/validation/rules.ts
@@ -1,6 +1,36 @@
-import { ValidationRule } from 'components/VForm/types';
-
 type Rule = (...params: any) => boolean | string;
+
+export enum ValidationRule {
+  After = 'after',
+  Alpha = 'alpha',
+  AlphaDash = 'alpha_dash',
+  AlphaNum = 'alpha_num',
+  AlphaSpaces = 'alpha_spaces',
+  Before = 'before',
+  Between = 'between',
+  Confirmed = 'confirmed',
+  CreditCard = 'credit_card',
+  DateBetween = 'date_between',
+  Decimal = 'decimal',
+  Digits = 'digits',
+  Image = 'image',
+  Included = 'included',
+  Integer = 'integer',
+  IP = 'ip',
+  IPOrFQDN = 'ip_or_fqdn',
+  Is = 'is',
+  IsNot = 'is_not',
+  Length = 'length',
+  Max = 'max',
+  Min = 'min',
+  Mimes = 'mimes',
+  Numeric = 'numeric',
+  Regex = 'regex',
+  Required = 'required',
+  RequiredIf = 'required_if',
+  Size = 'size',
+  URL = 'url'
+}
 
 export const rules: { [key: string]: Rule } = {
   [ValidationRule.After](value: string, referenceDate: string): boolean | string {
